+++ src/test/java/org/cicirello/permutations/PermutationTestCases.java
@@ -25,12 +25,11 @@ import java.util.SplittableRandom;
 import java.util.Random;
 import java.util.concurrent.ThreadLocalRandom;
 import java.math.BigInteger;
-import java.util.NoSuchElementException;
 import org.junit.jupiter.api.*;
 import static org.junit.jupiter.api.Assertions.*;
 
 /**
- * JUnit tests for the constructors and methods of the Permutation class, as well as the PermutationIterator.
+ * JUnit tests for the constructors and methods of the Permutation class.
  */
 public class PermutationTestCases {
 	
@@ -42,155 +41,6 @@ public class PermutationTestCases {
 	// Note: This doesn't disable those tests for the constructors/scramble methods
 	// that use a source of randomness that is seedable.
 	private static final boolean disableChiSquareTests = true;
-	
-	@Test
-	public void testIllegalPermutationStateExceptionConstructors() {
-		IllegalPermutationStateException e1 = new IllegalPermutationStateException("test message");
-		assertEquals("test message", e1.getMessage());
-		assertNull(e1.getCause());
-		e1 = new IllegalPermutationStateException("test message", new IllegalArgumentException());
-		assertEquals("test message", e1.getMessage());
-		assertTrue(e1.getCause() instanceof IllegalArgumentException);
-		e1 = new IllegalPermutationStateException("test message", new IllegalPermutationStateException("testing 1 2 3"));
-		assertEquals("test message", e1.getMessage());
-		assertTrue(e1.getCause() instanceof IllegalPermutationStateException);
-	}
-	
-	@Test
-	public void testUnaryOperator() {
-		Permutation p = new Permutation(10);
-		p.apply(perm -> { for (int i = 0; i < perm.length; i++) { perm[i] = i; }});
-		for (int i = 0; i < 10; i++) {
-			assertEquals(i, p.get(i));
-		}
-		p.apply(perm -> { for (int i = 0; i < perm.length; i++) { perm[perm.length-1-i] = i; }});
-		for (int i = 0; i < 10; i++) {
-			assertEquals(9-i, p.get(i));
-		}
-	}
-	
-	@Test
-	public void testValidatedUnaryOperator() {
-		final Permutation p = new Permutation(10);
-		p.applyThenValidate(perm -> { for (int i = 0; i < perm.length; i++) { perm[i] = i; }});
-		for (int i = 0; i < 10; i++) {
-			assertEquals(i, p.get(i));
-		}
-		p.applyThenValidate(perm -> { for (int i = 0; i < perm.length; i++) { perm[perm.length-1-i] = i; }});
-		for (int i = 0; i < 10; i++) {
-			assertEquals(9-i, p.get(i));
-		}
-		IllegalPermutationStateException thrown = assertThrows( 
-			IllegalPermutationStateException.class,
-			() -> p.applyThenValidate(perm -> { for (int i = 0; i < perm.length; i++) { perm[i] = 0; }})
-		);
-		assertTrue(thrown.getCause() instanceof IllegalArgumentException);
-	}
-	
-	@Test
-	public void testBinaryOperator() {
-		Permutation p1 = new Permutation(10);
-		Permutation p2 = new Permutation(10);
-		p1.apply((perm1, perm2) -> { for (int i = 0; i < perm1.length; i++) { perm1[i] = i; perm2[perm1.length-1-i] = i; }}, p2);
-		for (int i = 0; i < 10; i++) {
-			assertEquals(i, p1.get(i));
-			assertEquals(9-i, p2.get(i));
-		}
-		p1.apply((perm1, perm2) -> { for (int i = 0; i < perm1.length; i++) { perm2[i] = i; perm1[perm1.length-1-i] = i; }}, p2);
-		for (int i = 0; i < 10; i++) {
-			assertEquals(i, p2.get(i));
-			assertEquals(9-i, p1.get(i));
-		}
-	}
-	
-	@Test
-	public void testValidatedBinaryOperator() {
-		final Permutation p1 = new Permutation(10);
-		final Permutation p2 = new Permutation(10);
-		p1.applyThenValidate((perm1, perm2) -> { for (int i = 0; i < perm1.length; i++) { perm1[i] = i; perm2[perm1.length-1-i] = i; }}, p2);
-		for (int i = 0; i < 10; i++) {
-			assertEquals(i, p1.get(i));
-			assertEquals(9-i, p2.get(i));
-		}
-		p1.applyThenValidate((perm1, perm2) -> { for (int i = 0; i < perm1.length; i++) { perm2[i] = i; perm1[perm1.length-1-i] = i; }}, p2);
-		for (int i = 0; i < 10; i++) {
-			assertEquals(i, p2.get(i));
-			assertEquals(9-i, p1.get(i));
-		}
-		IllegalPermutationStateException thrown = assertThrows( 
-			IllegalPermutationStateException.class,
-			() -> p1.applyThenValidate((perm1, perm2) -> { for (int i = 0; i < perm1.length; i++) { perm1[i] = 0; perm2[perm1.length-1-i] = 0; }}, p2)
-		);
-		assertTrue(thrown.getCause() instanceof IllegalArgumentException);
-	}
-	
-	@Test
-	public void testFullUnaryOperator() {
-		Permutation p = new Permutation(10);
-		p.apply((perm, original) -> { for (int i = 0; i < perm.length; i++) { perm[i] = i; assertEquals(perm[i], original.get(i)); }});
-		for (int i = 0; i < 10; i++) {
-			assertEquals(i, p.get(i));
-		}
-		p.apply((perm, original) -> { for (int i = 0; i < perm.length; i++) { perm[perm.length-1-i] = i; assertEquals(perm[perm.length-1-i], original.get(perm.length-1-i)); }});
-		for (int i = 0; i < 10; i++) {
-			assertEquals(9-i, p.get(i));
-		}
-	}
-	
-	@Test
-	public void testValidatedFullUnaryOperator() {
-		final Permutation p = new Permutation(10);
-		p.applyThenValidate((perm, original) -> { for (int i = 0; i < perm.length; i++) { perm[i] = i; assertEquals(perm[i], original.get(i)); }});
-		for (int i = 0; i < 10; i++) {
-			assertEquals(i, p.get(i));
-		}
-		p.applyThenValidate((perm, original) -> { for (int i = 0; i < perm.length; i++) { perm[perm.length-1-i] = i; assertEquals(perm[perm.length-1-i], original.get(perm.length-1-i)); }});
-		for (int i = 0; i < 10; i++) {
-			assertEquals(9-i, p.get(i));
-		}
-		IllegalPermutationStateException thrown = assertThrows( 
-			IllegalPermutationStateException.class,
-			() -> p.applyThenValidate((perm, original) -> { for (int i = 0; i < perm.length; i++) { perm[i] = 0; }})
-		);
-		assertTrue(thrown.getCause() instanceof IllegalArgumentException);
-	}
-	
-	@Test
-	public void testFullBinaryOperator() {
-		Permutation p1 = new Permutation(10);
-		Permutation p2 = new Permutation(10);
-		p1.apply((perm1, perm2, o1, o2) -> { for (int i = 0; i < perm1.length; i++) { perm1[i] = i; perm2[perm1.length-1-i] = i; assertEquals(i, o1.get(i)); assertEquals(i, o2.get(9-i)); }}, p2);
-		for (int i = 0; i < 10; i++) {
-			assertEquals(i, p1.get(i));
-			assertEquals(9-i, p2.get(i));
-		}
-		p1.apply((perm1, perm2, o1, o2) -> { for (int i = 0; i < perm1.length; i++) { perm2[i] = i; perm1[perm1.length-1-i] = i; assertEquals(i, o2.get(i)); assertEquals(i, o1.get(9-i)); }}, p2);
-		for (int i = 0; i < 10; i++) {
-			assertEquals(i, p2.get(i));
-			assertEquals(9-i, p1.get(i));
-		}
-	}
-	
-	@Test
-	public void testValidatedFullBinaryOperator() {
-		final Permutation p1 = new Permutation(10);
-		final Permutation p2 = new Permutation(10);
-		p1.applyThenValidate((perm1, perm2, o1, o2) -> { for (int i = 0; i < perm1.length; i++) { perm1[i] = i; perm2[perm1.length-1-i] = i; assertEquals(i, o1.get(i)); assertEquals(i, o2.get(9-i)); }}, p2);
-		for (int i = 0; i < 10; i++) {
-			assertEquals(i, p1.get(i));
-			assertEquals(9-i, p2.get(i));
-		}
-		p1.applyThenValidate((perm1, perm2, o1, o2) -> { for (int i = 0; i < perm1.length; i++) { perm2[i] = i; perm1[perm1.length-1-i] = i; assertEquals(i, o2.get(i)); assertEquals(i, o1.get(9-i)); }}, p2);
-		for (int i = 0; i < 10; i++) {
-			assertEquals(i, p2.get(i));
-			assertEquals(9-i, p1.get(i));
-		}
-		IllegalPermutationStateException thrown = assertThrows( 
-			IllegalPermutationStateException.class,
-			() -> p1.applyThenValidate((perm1, perm2, o1, o2) -> { for (int i = 0; i < perm1.length; i++) { perm1[i] = 0; perm2[perm1.length-1-i] = 0; }}, p2)
-		);
-		assertTrue(thrown.getCause() instanceof IllegalArgumentException);
-	}
 	
 	@Test
 	public void testZeroLengthPermutations() {
@@ -1121,45 +971,6 @@ public class PermutationTestCases {
 		thrown = assertThrows( 
 			IllegalArgumentException.class,
 			() -> (new Permutation(10)).swapBlocks(1, 3, 5, 10)
-		);
-	}
-	
-	@Test
-	public void testPermutationIterator() {
-		int fact = 1;
-		for (int n = 1; n <= 5; n++) {
-			Permutation p = new Permutation(n);
-			boolean checkedFirst = false;
-			fact *= n;
-			boolean[] found = new boolean[fact];
-			int count = 0;
-			for (Permutation pPrime : p) {
-				if (!checkedFirst) {
-					assertEquals(p, pPrime);
-					checkedFirst = true;
-				}
-				int permID = pPrime.toInteger();
-				assertFalse(found[permID]);
-				found[permID] = true;
-				count++;
-			}
-			assertEquals(fact, count);
-		}
-		final PermutationIterator iter = new PermutationIterator(4);
-		fact = 24;
-		boolean[] found = new boolean[fact];
-		int count = 0;
-		while (iter.hasNext()) {
-			Permutation p = iter.next();
-			int permID = p.toInteger();
-			assertFalse(found[permID]);
-			found[permID] = true;
-			count++;
-		}
-		assertEquals(fact, count);
-		NoSuchElementException thrown = assertThrows( 
-			NoSuchElementException.class,
-			() -> iter.next()
 		);
 	}
